Closes #1272, image loading has been replaced for a less custom solution that works better and shows loading progress if a lot of places (post cells, pins, history, album cells)
Also related to #1135, #1126 (gradle dep for Material has been updated, tell me if things are wonky)
Closes #1318, flags on catalog mode posts are now positioned better so they're always visible
Closes #1284, the image loader interface will load a thumbnail before loading a high quality image, and if a high-quality load fails, the low quality one will remain
Closes #1331 again, but MUCH more consistently
Closes #1279, I believe this was much like the thumbnail issues as well
Closes #1315, staggering has been added for the album seperate from the catalog
Also album cell info has been removed and the overflow menu now has a "image info" option that mirrors the post info if you REALLY want to know about an image
A number of layouts using ConstraintLayout have been replaced with lighter layouts where applicable
PostIcons has been more properly implemented, as it had issues with measurement and XML layout; it's API has also been cleaned up
Changing to the stagger mode for catalogs will now properly update the layout if being viewed
The post popup replies container has been generified so that modifying the buttons applies to all layouts using it
The removed posts container and layouts have been refreshed to be more palatable
The theme settings controller now allows for images to be tapped, and test snackbars and toasts to be shown, as well as some extra span stuff
A number of classes have been removed in preference to existing platform classes that do the job better, or extra inheritance where it was not needed
Filters can now have partially transparent highlights; the associated ColorPickerView has been updated much the same way as PostIcons was

+++ Kuroba/app/src/main/java/com/github/adamantcheese/chan/core/net/NetUtils.java
@@ -246,21 +246,21 @@ public class NetUtils {
     public static Call makeBitmapRequest(
             final HttpUrl url, @NonNull final BitmapResult result
     ) {
-        return makeBitmapRequest(url, result, 0, 0);
+        return makeBitmapRequest(url, result, 0, 0, null);
     }
 
     /**
      * Request a bitmap without resizing, with timeout
      *
-     * @param url    The request URL. If null, null will be returned.
-     * @param result The callback for this call.
-     * @param timeoutMs  Optional timeout value, in milliseconds (-1 for no timeout)
+     * @param url       The request URL. If null, null will be returned.
+     * @param result    The callback for this call.
+     * @param timeoutMs Optional timeout value, in milliseconds (-1 for no timeout)
      * @return An enqueued bitmap call. WILL RUN RESULT ON MAIN THREAD!
      */
     public static Call makeBitmapRequest(
             final HttpUrl url, @NonNull final BitmapResult result, final int timeoutMs
     ) {
-        Pair<Call, Callback> ret =  makeBitmapRequest(url, result, 0, 0, timeoutMs, true, true);
+        Pair<Call, Callback> ret = makeBitmapRequest(url, result, 0, 0, timeoutMs, null, true, true);
         return ret == null ? null : ret.first;
     }
 
@@ -274,9 +274,13 @@ public class NetUtils {
      * @return An enqueued bitmap call. WILL RUN RESULT ON MAIN THREAD!
      */
     public static Call makeBitmapRequest(
-            final HttpUrl url, @NonNull final BitmapResult result, final int width, final int height
+            final HttpUrl url,
+            @NonNull final BitmapResult result,
+            final int width,
+            final int height,
+            @Nullable ProgressResponseBody.ProgressListener progressListener
     ) {
-        Pair<Call, Callback> ret = makeBitmapRequest(url, result, width, height, -1, true, true);
+        Pair<Call, Callback> ret = makeBitmapRequest(url, result, width, height, -1, progressListener, true, true);
         return ret == null ? null : ret.first;
     }
 
@@ -298,6 +302,7 @@ public class NetUtils {
             final int width,
             final int height,
             final int timeoutMs,
+            @Nullable final ProgressResponseBody.ProgressListener progressListener,
             boolean enqueue,
             boolean mainThread
     ) {
@@ -322,11 +327,16 @@ public class NetUtils {
         if (timeoutMs != -1) {
             client = client.newBuilder().callTimeout(timeoutMs, TimeUnit.MILLISECONDS).build();
         }
-        Call call = client
-                .newCall(new Request.Builder().url(url)
-                        .addHeader("Referer", url.toString())
-                        .cacheControl(new CacheControl.Builder().maxStale(365, TimeUnit.DAYS).build())
-                        .build());
+        client = client.newBuilder().addNetworkInterceptor(chain -> {
+            Response originalResponse = chain.proceed(chain.request());
+            return originalResponse.newBuilder()
+                    .body(new ProgressResponseBody(originalResponse, progressListener))
+                    .build();
+        }).build();
+        Call call = client.newCall(new Request.Builder().url(url)
+                .addHeader("Referer", url.toString())
+                .cacheControl(new CacheControl.Builder().maxStale(365, TimeUnit.DAYS).build())
+                .build());
         Callback callback = new Callback() {
             @Override
             public void onFailure(@NotNull Call call, @NotNull IOException e) {
@@ -363,11 +373,17 @@ public class NetUtils {
                             );
                         }
                         FilesKt.writeBytes(tempFile, body.bytes());
-                        BitmapUtils.decodeFilePreviewImage(tempFile, 0, 0, bitmap -> {
+                        Bitmap result =
+                                BitmapUtils.decodeFilePreviewImage(tempFile, 0, 0, mainThread ? null : bitmap -> {
+                                    //noinspection ResultOfMethodCallIgnored
+                                    tempFile.delete();
+                                    checkBitmap(url, bitmap, false);
+                                }, false);
+                        if (result != null) {
                             //noinspection ResultOfMethodCallIgnored
                             tempFile.delete();
-                            checkBitmap(url, bitmap, mainThread);
-                        }, false);
+                            checkBitmap(url, result, true);
+                        }
                     } else {
                         ExceptionCatchingInputStream wrappedStream =
                                 new ExceptionCatchingInputStream(body.byteStream());
@@ -418,7 +434,7 @@ public class NetUtils {
 
     private static synchronized void performBitmapFailure(@NonNull final HttpUrl url, Exception e, boolean mainThread) {
         final List<BitmapResult> results = resultListeners.remove(url);
-        if (results == null) return;
+        if (results == null || e instanceof StreamResetException || "Canceled".equals(e.getMessage())) return;
         for (final BitmapResult bitmapResult : results) {
             if (bitmapResult == null) continue;
             if (mainThread) {
